adressed more comments
