+++ 2020-spring/Kraynov Alexander/hw2/Mtg Cards/Mtg Cards/Service/CardFacade.swift
@@ -7,45 +7,10 @@
 //
 
 import Foundation
-import RealmSwift
-//swiftlint:disable:next file_types_order
+
 protocol CardFacade {
     typealias OnUpdateCompletion = ([Card]?) -> Void
 
     func getCards(completion: @escaping OnUpdateCompletion)
     func loadMore(completion: @escaping OnUpdateCompletion)
-}
-
-final class CardFacadeImpl: CardFacade {
-    private let cardRepository: CardRepository
-    private let cardService: CardService
-    private var cardToken: NotificationToken?
-
-    init(cardService: CardService, cardRepository: CardRepository) {
-        self.cardRepository = cardRepository
-        self.cardService = cardService
-    }
-
-    func getCards(completion: @escaping OnUpdateCompletion) {
-        cardService.getCards { cards in
-            guard let cards = cards else {
-                return
-            }
-            self.cardRepository.save(cards)
-        }
-        let cards = cardRepository.getCards()
-        cardToken = cards.observe { _ in
-            completion(cards.map { $0.card })
-        }
-    }
-
-    func loadMore(completion: @escaping OnUpdateCompletion) {
-        cardService.getMoreCards { cards in
-            guard let cards = cards else {
-                return
-            }
-            self.cardRepository.save(cards)
-            completion(Array(cards))
-        }
-    }
 }
