+++ 2020-spring/Kraynov Alexander/hw2/Mtg Cards/Mtg Cards/Service/CardFacadeImpl.swift
@@ -1,0 +1,44 @@
+//
+//  CardFacadeImpl.swift
+//  Mtg Cards
+//
+//  Created by Admin on 31.03.2020.
+//  Copyright © 2020 alexander. All rights reserved.
+//
+
+import Foundation
+import RealmSwift
+
+final class CardFacadeImpl: CardFacade {
+    private let cardRepository: CardRepository
+    private let cardService: CardService
+    private var cardToken: NotificationToken?
+
+    init(cardService: CardService, cardRepository: CardRepository) {
+        self.cardRepository = cardRepository
+        self.cardService = cardService
+    }
+
+    func getCards(completion: @escaping OnUpdateCompletion) {
+        cardService.getCards { cards in
+            guard let cards = cards else {
+                return
+            }
+            self.cardRepository.save(cards)
+        }
+        let cards = cardRepository.getCards()
+        cardToken = cards.observe { _ in
+            completion(cards.map { $0.card })
+        }
+    }
+
+    func loadMore(completion: @escaping OnUpdateCompletion) {
+        cardService.getMoreCards { cards in
+            guard let cards = cards else {
+                return
+            }
+            self.cardRepository.save(cards)
+            completion(Array(cards))
+        }
+    }
+}
