+++ 2020-spring/Kraynov Alexander/hw2/Mtg Cards/Mtg Cards/Service/CardRepository.swift
@@ -9,28 +9,7 @@
 import Foundation
 import RealmSwift
 
-//swiftlint:disable:next file_types_order
 protocol CardRepository {
     func save (_ characters: [Card])
     func getCards() -> Results<CardRealm>
-}
-
-final class CardRepositoryImpl: CardRepository {
-    var realm: Realm {
-        do {
-            return try Realm()
-        } catch {
-            fatalError("Realm can't be created")
-        }
-    }
-
-    func save(_ cards: [Card]) {
-        let cards = cards.map(CardRealm.init(card: ))
-        try? realm.write {
-            realm.add(cards, update: .modified)
-        }
-    }
-    func getCards() -> Results<CardRealm> {
-        realm.objects(CardRealm.self)
-    }
 }
