+++ 2020-spring/Kraynov Alexander/hw2/Mtg Cards/Mtg Cards/Service/CardServiceImpls.swift
@@ -1,0 +1,9 @@
+//
+//  CardServiceImpls.swift
+//  Mtg Cards
+//
+//  Created by Admin on 31.03.2020.
+//  Copyright © 2020 alexander. All rights reserved.
+//
+
+import Foundation
